refactor(zed): extract sdkless ZedRecorder base; document macOS IMU rate quirk
ZedMid360Recorder now just adds lidar/odometry ports on top of the reusable
ZedRecorder (stereo h264 + camera_info + batched 800 Hz IMU path). The
sdkless_imu docstring documents the Apple Silicon root-port ~533 Hz cap and
the USB-hub workaround (verified 795-800 Hz).
